refactor: Modifier.drawUnderBarWithAnimation 위치 이동
- TextFieidImpl -> draw
